Add format check on device chatacteristics test and add the track event to the run query handler
