+++ src/telemetry/telemetry.ts
@@ -29,6 +29,7 @@ import {
 } from './filters';
 import ITelemetry from './ITelemetry';
 import { getVersion } from '../app/utils/version';
+import { getBrowserScreenSize, getDeviceScreenScale } from '../app/utils/device-characteristics-telemetry';
 
 class Telemetry implements ITelemetry {
   private appInsights: ApplicationInsights;
@@ -119,9 +120,19 @@ class Telemetry implements ITelemetry {
     telemetry.trackEvent(WINDOW_OPEN_EVENT, properties);
   }
 
-  public trackDeviceCharacteristics(properties?: any){
-    properties = properties || {}
+  public trackDeviceCharacteristicsTelemetry(properties?: any){
+    const deviceProperties = {
+      deviceHeight: screen.height,
+      deviceWidth: screen.width,
+      browserScreenSize: getBrowserScreenSize(window.innerWidth),
+      browserHeight: window.innerHeight,
+      browserWidth: window.innerWidth,
+      scale: getDeviceScreenScale()
+    };
+    properties = properties || deviceProperties;
+
     telemetry.trackEvent(DEVICE_CHARACTERISTICS_EVENT, properties);
+    console.log('telemetry collected');
   }
 
   private getInstrumentationKey() {
